Add state count tests for mf6 examples
The number of start and end states of stress period, time step, and
iteration needs to match for each model. This checks for the counts.
It does not look at the sequence. Furthermore, five models seem to have
other problems with the current and total number of time steps and the
condition `sim_grp.nstp == sim_grp.kstp + 1` in line 113 in `runner.py`
is not met as often as it should. This is a working hypothesis that
requires deeper investigation. Skip the count tests for these models for
now.